fix(server): prevent decorator arg ordering from preventing handler lookup
When passing args through the message handler decorator, the
microservice server's default implementation for looking up handlers
involved a step where the keys of an object would be sorted, which
resulted in the possibility of different sorting orders. We've
overridden the method that does this sorting, which simplifies all of
the logic in general.