ci(tsdoc): deploy documentation to github pages
`gh-pages` is a small npm package which automatically deploys a
directory to github pages by commiting it to a branch called `gh-pages`.
We need to specify the user to hopefully support the wonky git repo
settings in Google Cloudbuild. If this doesn't work because of GCB
limitations we'll have to find another route to a solution.

resolves #25