feat(strategy): make subscription names always pass through strategy
To ensure consistency in behavior, the provided subscription name will
always be passed through the subscription naming strategy. I set up a
discriminated union to ensure completeness of handling. An exception is
thrown if the topic and subscription names are both not provided.

BREAKING CHANGE:
`SubscriptionNamingStrategy`'s interface has changed such that it
accepts an object. See BasicSubscriptionNamingStrategy as an example.